Squashed 'cloudburst/' changes from 4714e5d..9774b51
9774b51 Merge remote-tracking branch 'origin/master'
437d213 merged changes from dev
dd77164 merged changes from dev
b0d7b41 adjust to recent availability zone change in aws

git-subtree-dir: cloudburst
git-subtree-split: 9774b512299ea71a23e00029c0a54406f24c7b81

+++ scripts/fwlib.py
@@ -20,7 +20,7 @@ EXIT_ON_ERROR = True
 
 
 # get the input data for all the sites we are going to process
-def process_fetches(mode: str, tmp_zip_dir: str, fetches: list, prior_errors:bool = False):
+def process_fetches(mode: str, tmp_zip_dir: str, fetches: list, prior_errors: bool = False):
     return_code = 0
 
     # TODO: move elsewhere or change default
@@ -72,17 +72,29 @@ def process_fetches(mode: str, tmp_zip_dir: str, fetches: list, prior_errors:boo
             break
         elif expand:
             # unzip the inputs, prepare input files
-            expand_to = Path(dest).parent
-            
-            return_code = ziplib.expand(dest, expand_to)
-            if exit_on_error and return_code != 0:
-                break
-            elif remove_zips:
-                if len(dest) > 3 and dest[-3:] == '.7z':
-                    os.remove(dest)
-                elif dest.endswith('/'):
-                    for file in Path(dest).glob('*.7z'):
-                        file.unlink()
+            is_expanded = False
+            if dest.endswith('.7z'):
+                expand_to = Path(dest).parent
+            else:
+                # if dest does not end with 7z, it may be a directory tree with 7z files to expand
+                p_dest = Path(dest)
+                if p_dest.is_dir():
+                    for zip in p_dest.rglob('*.7z'):
+                        return_code = ziplib.expand(zip, zip.parent)
+                        if exit_on_error and return_code != 0:
+                            break
+                        elif remove_zips:
+                            zip.unlink()
+                    is_expanded = True
+
+            if not is_expanded:
+                return_code = ziplib.expand(dest, expand_to)
+
+                if exit_on_error and return_code != 0:
+                    break
+                elif remove_zips:
+                    if len(dest) > 3 and dest[-3:] == '.7z':
+                        os.remove(dest)
 
         if 'excludeFilePattern' in fetch:
             patterns = fetch['excludeFilePattern']
@@ -94,7 +106,7 @@ def process_fetches(mode: str, tmp_zip_dir: str, fetches: list, prior_errors:boo
     return return_code
 
 
-def run_tasks(process_list: list, mode: str, prior_errors:bool = False):
+def run_tasks(process_list: list, mode: str, prior_errors: bool = False):
     logFolder = "./logs"
     return_code = 0
     
@@ -253,7 +265,7 @@ def manage_process(process_name: str, work_list: list, command: str, log_behavio
     return return_code
 
 
-def move_files(move_tasks: list, mode: str, prior_errors:bool = False):
+def move_files(move_tasks: list, mode: str, prior_errors: bool = False):
     return_code = 0
     
     for task in move_tasks:
@@ -311,7 +323,7 @@ def move_files(move_tasks: list, mode: str, prior_errors:bool = False):
     return return_code
 
 
-def process_store(tasks_store: list, mode: str, prior_errors:bool = False):
+def process_store(tasks_store: list, mode: str, prior_errors: bool = False):
     return_code = 0
 
     for task in tasks_store:
@@ -347,9 +359,9 @@ def process_store(tasks_store: list, mode: str, prior_errors:bool = False):
         if 'compress' in task:
             compress = task['compress']
 
-        compressSubDirectories = False
+        compress_sub_directories = False
         if 'compressSubDirectories' in task:
-            compressSubDirectories = task['compressSubDirectories']
+            compress_sub_directories = task['compressSubDirectories']
 
         remove_on_store = True
         if 'removeOnStore' in task:
@@ -365,7 +377,7 @@ def process_store(tasks_store: list, mode: str, prior_errors:bool = False):
             is_empty = False
             if Path(source).is_dir():
                 contents = list(Path(source).glob('*'))
-                if not compressSubDirectories:
+                if not compress_sub_directories:
                     # is there anything to zip?
                     is_empty = (len(contents) == 0)
                 else:
@@ -380,7 +392,7 @@ def process_store(tasks_store: list, mode: str, prior_errors:bool = False):
             else:
                 print(f"saving outputs: {source}")
 
-                if compressSubDirectories and len(contents) > 0:
+                if compress_sub_directories and len(contents) > 0:
                     for dir in contents:
                         if dir.is_dir():
                             src = str(dir)
